Create useEntries custom hook to separate the fetching logic from the component
